+++ client/src/actions/index.js
@@ -1,10 +1,14 @@
 import axios from "axios";
-import { FETCH_USER, LOGOUT_USER } from "./types";
+import { FETCH_USER, LOGOUT_USER, LOGIN_USER, FETCH_SURVEYS } from "./types";
 
 export const fetchUser = () => async dispatch => {
   const res = await axios.get("/api/current_user");
-
   dispatch({ type: FETCH_USER, payload: res.data });
+};
+
+export const loginUser = user => async dispatch => {
+  const res = await axios.post("/api/login", user);
+  dispatch({ type: LOGIN_USER, payload: res.data });
 };
 
 export const logOutUser = () => async dispatch => {
@@ -17,4 +21,16 @@ export const handleToken = token => async dispatch => {
   const res = await axios.post("/api/stripe", token);
 
   dispatch({ type: FETCH_USER, payload: res.data });
+};
+
+export const submitSurvey = (values, history) => async dispatch => {
+  const res = await axios.post("/api/surveys", values);
+  history.push("/surveys");
+  dispatch({ type: FETCH_USER, payload: res.data });
+};
+
+export const fetchSurveys = () => async dispatch => {
+  const res = await axios.get("/api/surveys");
+
+  dispatch({ type: FETCH_SURVEYS, payload: res.data });
 };