+++ client/src/actions/types.js
@@ -1,2 +1,4 @@
 export const FETCH_USER = "FETCH_USER";
+export const FETCH_SURVEYS = "FETCH_SURVEYS";
+export const LOGIN_USER = "LOGIN_USER";
 export const LOGOUT_USER = "LOGOUT_USER";