+++ client/src/index.js
@@ -10,6 +10,9 @@ import { Provider } from "react-redux";
 import thunk from "redux-thunk";
 import reducers from "./reducers";
 
+import axios from "axios";
+window.axios = axios;
+
 const store = createStore(reducers, {}, applyMiddleware(thunk));
 
 ReactDOM.render(