+++ model/User.js
@@ -1,13 +1,13 @@
 const mongoose = require("mongoose");
 
-const { Schema } = mongoose;
-
-const userSchema = new Schema({
+const userSchema = new mongoose.Schema({
   googleId: String,
   credits: {
     type: Number,
     default: 0
-  }
+  },
+  email: String,
+  password: String
 });
 
 mongoose.model("users", userSchema);