+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change log
 
+## v1.0.0 29 Aug, 2019
+
+#### All Email Survey Functionality
+
+**Added Backend:**
+
+- middleware/requireCredits.js
+- model/Recipient.js
+- model/Survey.js
+- routes/surveyRoutes.js
+- sendgrid_webhook.js
+- services/Mailer.js
+- services/emailTemplates/
+
+**Added FrontEnd:**
+
+- client/public/img/
+- client/src/components/Dashboard.jsx
+- client/src/components/auth/
+- client/src/components/surveys/
+- client/src/img/
+- client/src/reducers/surveysReducer.js
+- client/src/utils/
+
 ## v1.0.0 1 Aug, 2019
 
 #### Client side added with create-react-app